requirements.txt
telethon
python-dotenv
openai
